% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM_algo}
\alias{EM_algo}
\title{EM_algo: Compute the EM algorithm in a dataset}
\usage{
EM_algo(data, K, epsilon, type_init = "kmeans", parameters = 0)
}
\arguments{
\item{data}{: main dataset that will be used for the algorithm}

\item{K}{: number of clusters}

\item{epsilon}{: stop criteria for a loss decrease lower then this value}

\item{type_init}{: choose parameter initialization.
There are three options: \link{"random", "kmeans"}.
"random" is random initialization, "kmeans" is with the kmeans algo}

\item{parameters}{: model parameters}
}
\value{
: log_likelihood : log likelihood values of the iterations
y_pred : cluster label of each data point
maxll: maximum log likelihood
parameters: the model parameters (mean, variance, tk etc)
}
\description{
Implementation of the EM algorithm for the use for a Mixture Model
problem
}
\examples{
 K <- 5
split <- sample(1:150, 100)
y <- iris[-split, 5]
data <- iris[-split, -5]
epsilon <- 10^-6
data <- as.matrix(data)

res <- EM_algo(data, K, epsilon, "random")
}
